tipousuario

+++ src/main/java/com/ausiasmarch/deckrift/service/UsuarioService.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 
 import com.ausiasmarch.deckrift.entity.UsuarioEntity;
 import com.ausiasmarch.deckrift.exception.ResourceNotFoundException;
+import com.ausiasmarch.deckrift.exception.UnauthorizedAccessException;
 import com.ausiasmarch.deckrift.repository.UsuarioRepository;
 
 @Service
@@ -20,6 +21,18 @@ public class UsuarioService implements ServiceInterface<UsuarioEntity> {
     @Autowired
     RandomService oRandomService;
 
+    AuthService oAuthService;
+
+    public UsuarioEntity getByEmail(String email) {
+        UsuarioEntity oUsuarioEntity = oUsuarioRepository.findByCorreo(email)
+                .orElseThrow(() -> new ResourceNotFoundException("El usuario con email " + email + " no existe"));
+        if (oAuthService.isAdmin() || oAuthService.isAuditorWithItsOwnData(oUsuarioEntity.getId())) {
+            return oUsuarioEntity;
+        } else {
+            throw new UnauthorizedAccessException("No tienes permisos para ver el usuario");
+        }
+    }
+
     // Obtener un usuario por ID
     public UsuarioEntity findById(Long id) {
         return oUsuarioRepository.findById(id)
@@ -27,12 +40,16 @@ public class UsuarioService implements ServiceInterface<UsuarioEntity> {
     }
 
     public Page<UsuarioEntity> getPage(Pageable oPageable, Optional<String> filter) {
-
-        if (filter.isPresent()) {
-            return oUsuarioRepository
-                    .findByNombreContaining(filter.get(), oPageable);
+        if (oAuthService.isAdmin()) {
+            if (filter.isPresent()) {
+                return oUsuarioRepository
+                        .findByNombreContaining(
+                                filter.get(), oPageable);
+            } else {
+                return oUsuarioRepository.findAll(oPageable);
+            }
         } else {
-            return oUsuarioRepository.findAll(oPageable);
+            throw new UnauthorizedAccessException("No tienes permisos para ver los usuarios");
         }
     }
 
@@ -43,7 +60,11 @@ public class UsuarioService implements ServiceInterface<UsuarioEntity> {
     }
 
     public Long count() {
-        return oUsuarioRepository.count();
+        if (!oAuthService.isAdmin()) {
+            throw new UnauthorizedAccessException("No tienes permisos para contar los usuarios");
+        } else {
+            return oUsuarioRepository.count();
+        }
     }
 
     public UsuarioEntity randomSelection() {
@@ -52,32 +73,40 @@ public class UsuarioService implements ServiceInterface<UsuarioEntity> {
 
     // Crear un nuevo usuario
     public UsuarioEntity create(UsuarioEntity oUsuarioEntity) {
-        return oUsuarioRepository.save(oUsuarioEntity);
+            return oUsuarioRepository.save(oUsuarioEntity);
     }
 
     // Actualizar un usuario existente
     public UsuarioEntity update(UsuarioEntity oUsuarioEntity) {
-        UsuarioEntity oUsuarioEntityFromDatabase = oUsuarioRepository.findById(oUsuarioEntity.getId()).get();
-        if (oUsuarioEntity.getNombre() != null) {
-            oUsuarioEntityFromDatabase.setNombre(oUsuarioEntity.getNombre());
+        if (oAuthService.isAdmin() || oAuthService.isAuditorWithItsOwnData(oUsuarioEntity.getId())) {
+            UsuarioEntity oUsuarioEntityFromDatabase = oUsuarioRepository.findById(oUsuarioEntity.getId()).get();
+            if (oUsuarioEntity.getNombre() != null) {
+                oUsuarioEntityFromDatabase.setNombre(oUsuarioEntity.getNombre());
+            }
+            if (oUsuarioEntity.getCorreo() != null) {
+                oUsuarioEntityFromDatabase.setCorreo(oUsuarioEntity.getCorreo());
+            }
+            if (oUsuarioEntity.getPassword() != null) {
+                oUsuarioEntityFromDatabase.setPassword(oUsuarioEntity.getPassword());
+            }
+            return oUsuarioRepository.save(oUsuarioEntityFromDatabase);
+        } else {
+            throw new UnauthorizedAccessException("No tienes permisos para modificar el usuario");
         }
-        if (oUsuarioEntity.getCorreo() != null) {
-            oUsuarioEntityFromDatabase.setCorreo(oUsuarioEntity.getCorreo());
-        }
-        if (oUsuarioEntity.getPassword() != null) {
-            oUsuarioEntityFromDatabase.setPassword(oUsuarioEntity.getPassword());
-        }
-        return oUsuarioRepository.save(oUsuarioEntityFromDatabase);
     }
 
     // Eliminar un usuario por ID
     public Long delete(Long id) {
-        oUsuarioRepository.deleteById(id);
-        return 1L;
+            oUsuarioRepository.deleteById(id);
+            return 1L;
     }
 
     public Long deleteAll() {
-        oUsuarioRepository.deleteAll();
-        return this.count();
+        if (!oAuthService.isAdmin()) {
+            throw new UnauthorizedAccessException("No tienes permisos para borrar todos los usuarios");
+        } else {
+            oUsuarioRepository.deleteAll();
+            return this.count();
+        }
     }
 }
